requirements.txt
streamlit
gspread
gspread-formatting
google-auth
pandas